Added new features
Now Generates a list off all possible move the player (AI) can do and inputs them into BadAI(). That now returns a value that looks like [[x,y],[x,y]] I then take that input and put it into InputToMove() that moves it. This streamlines the process so that when we implement the real AI it will work better I think.

THERE IS A NEW ERROR ON GAME END THAT MAKES IT CRASH :)